Add styles to headers and text elements
Created and linked an external stylesheet for the purpose of starting to style the recipe pages as well as the index.html file making changes to the headers, paragraph and list elements. The changes made include changing the font family and colours that used across the pages; changing the backround colour of the headers to help them stand out and adding margin to list elements to make them more readable.